+++ testAngularAPI.Server/Model/User.cs
@@ -1,0 +1,25 @@
+using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
+
+namespace testAngularAPI.Server.Model
+{
+    [BsonIgnoreExtraElements]
+    public class User
+    {
+        [BsonId]
+        [BsonRepresentation(BsonType.ObjectId)]
+        public string? Id { get; set; }
+
+        [BsonElement("Name")]
+        public string Name { get; set; }
+
+        [BsonElement("Email")]
+        public string Email { get; set; }
+
+        public User(string name, string email)
+        {
+            Name = name;
+            Email = email;
+        }
+    }
+}